Custom UI for the default system-managed composite evaluation
